--- src/_varibalse.scss
@@ -1,3 +1,0 @@
-$primary: #0000FF;
-$white: #f2f2f2;
-$dark: #212121;
