+++ src/index.js
@@ -7,6 +7,7 @@ import { Route, Switch } from 'react-router-dom';
 import { ConnectedRouter } from 'react-router-redux';
 
 import App from './components/App';
+import './index.scss'
 
 ReactDOM.render((
   <Provider store={store}>
