+++ src/index.scss
@@ -1,0 +1,18 @@
+:root{
+  --primary:#0000FF;
+  --white:#f2f2f2;
+  --dark:#212121;
+  --dark-text:#0A0A0B;
+  --gray-text:#62626A;
+  --red-text:#F53D5C;
+}
+@font-face {
+  font-family: 'LucidaSans';
+  src: url('./assets/fonts/Lucida-Sans-Unicode.woff') format('woff'),
+      url('./assets/fonts/Lucida-Sans-Unicode.woff2') format('woff2');
+  font-weight: 400;
+}
+
+body{
+  font-family: 'LucidaSans', sans-serif;
+}
